moving code to pages from components
