+++ client/src/Components/AdminInstructorTable.js
@@ -1,26 +1,8 @@
-import React, { useState, useEffect } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import React, { useState } from 'react'
 import { Table } from "react-bootstrap";
 import "../instructor.css";
-import { getAllInstructors } from '../actions'
 
 export default function AdminInstructorTable() {
-  const [instructorList, setInstructor] = useState();
-
-  const dispatch = useDispatch()
-
-  const [allInstructors, error, authObj] = useSelector((gState) => [
-    gState.allInstructors,
-    gState.error,
-    gState.authObj
-  ])
-
-  useEffect(() => {
-    dispatch(getAllInstructors(authObj.accessToken))
-  }, [])
-
-  console.log(allInstructors)
-  console.log(allInstructors.data)
 
   return (
     <Table className="table-expand">
